backend/requirements.txt
aiocoap==0.4.7
aiohttp==3.8.4
aiosignal==1.3.1
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==23.1.0
billiard==4.1.0
boto3==1.26.153
botocore==1.29.153
celery==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
cryptography==40.0.2
defusedxml==0.7.1
Django==4.2
django-celery-results==2.5.1
django-cors-headers==3.14.0
django-redis==5.2.0
django-storages==1.13.2
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
drf-extra-fields==3.4.1
frozenlist==1.3.3
gevent==22.10.2
greenlet==2.0.2
gunicorn==20.1.0
idna==3.4
jmespath==1.0.1
kombu==5.3.0
MarkupSafe==2.1.3
multidict==6.0.4
oauthlib==3.2.2
Pillow==9.5.0
prompt-toolkit==3.0.38
psycopg2==2.9.6
pycparser==2.21
PyJWT==2.6.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3
redis==4.5.5
requests==2.28.2
requests-oauthlib==1.3.1
s3transfer==0.6.1
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
tzdata==2023.3
urllib3==1.26.15
vine==5.0.0
wcwidth==0.2.6
Werkzeug==2.3.6
yarl==1.9.2
zope.event==4.6
zope.interface==6.0
